adding versioned sidebar component
