+++ app/components/layout/sidebar/v1/component.html.erb
@@ -141,7 +141,7 @@
         <ul
           id="sidebar-dd-menu"
           data-dropdown--v1-target="menu"
-          class="z-20 mx-3 hidden w-56 divide-y divide-slate-100 rounded-lg bg-white shadow-sm dark:bg-slate-700"
+          class="z-20 divide-y divide-slate-100 rounded-lg bg-white shadow-sm dark:bg-slate-700"
           tabindex="-1"
           hidden
           role="menu"
