+++ config/features.yml
@@ -42,6 +42,9 @@ features:
   v2_select2:
     description: "Enable the new v2 select2 component in the UI."
     enable_in_development: true
+  v2_sidebar:
+    description: "Enable the new v2 sidebar component in the UI."
+    enable_in_development: true
   wes_extended_metadata:
     description: "Include additional metadata in WES submission, such as namespaceId\
       \ and samplesCount."
